Completed all questions from part 4.

+++ bloglist/controllers/bloglist.js
@@ -1,7 +1,6 @@
 require('express-async-errors');
 const bloglistRouter = require('express').Router();
 const Blog = require('../models/blog');
-const User = require('../models/user');
 
 bloglistRouter.get('/', async (req, res) => {
 	const data = await Blog.find({}).populate('user', { name: 1, username: 1 });
@@ -10,8 +9,7 @@ bloglistRouter.get('/', async (req, res) => {
 
 bloglistRouter.post('/', async (req, res) => {
 	const body = req.body;
-	const user = await User.findById(body.user);
-
+	const user = req.user;
 	const newBlog = new Blog({
 		...body,
 		user: user._id
@@ -19,14 +17,20 @@ bloglistRouter.post('/', async (req, res) => {
 
 	const response = await newBlog.save();
 
-	user.blogs = user.blogs.concat(response._id);
-	await user.save();
+	req.user.blogs = req.user.blogs.concat(response._id);
+	await req.user.save();
 	res.status(201).json(response);
 });
 
 bloglistRouter.delete('/:id', async (req, res) => {
-	await Blog.findByIdAndDelete(req.params.id);
-	res.status(204).end();
+	const blog = await Blog.findById(req.params.id);
+	const user = req.user;
+
+	if (user.id.toString() === blog.user.toString()) {
+		await Blog.findByIdAndDelete(req.params.id);
+		res.status(204).end();
+	}
+	res.status(400).json({ error: 'Not the owner of the blog' });
 });
 
 bloglistRouter.put('/:id', async (req, res) => {
